Consolidate parallel WSGI implementations
