+++ beeline/middleware/bottle/__init__.py
@@ -1,6 +1,6 @@
 import beeline
 from beeline.propagation import Request
-from beeline.middleware.werkzeug import WerkzeugRequest
+from beeline.middleware.wsgi import WSGIRequest
 
 
 class HoneyWSGIMiddleware(object):
@@ -9,9 +9,9 @@ class HoneyWSGIMiddleware(object):
         self.app = app
 
     def __call__(self, environ, start_response):
-        wr = WerkzeugRequest(environ)
-        request_context = self.get_context_from_environ(environ)
-        root_span = beeline.propagate_and_start_trace(request_context, wr)
+        wr = WSGIRequest("bottle", environ)
+
+        root_span = beeline.propagate_and_start_trace(wr.request_context(), wr)
 
         def _start_response(status, headers, *args):
             beeline.add_context_field("response.status_code", status)
@@ -20,23 +20,3 @@ class HoneyWSGIMiddleware(object):
             return start_response(status, headers, *args)
 
         return self.app(environ, _start_response)
-
-    def get_context_from_environ(self, environ):
-        request_method = environ.get('REQUEST_METHOD')
-        if request_method:
-            trace_name = "bottle_http_%s" % request_method.lower()
-        else:
-            trace_name = "bottle_http"
-
-        return {
-            "name": trace_name,
-            "type": "http_server",
-            "request.host": environ.get('HTTP_HOST'),
-            "request.method": request_method,
-            "request.path": environ.get('PATH_INFO'),
-            "request.remote_addr": environ.get('REMOTE_ADDR'),
-            "request.content_length": environ.get('CONTENT_LENGTH', 0),
-            "request.user_agent": environ.get('HTTP_USER_AGENT'),
-            "request.scheme": environ.get('wsgi.url_scheme'),
-            "request.query": environ.get('QUERY_STRING')
-        }
