+++ beeline/middleware/flask/__init__.py
@@ -7,7 +7,7 @@ from beeline.propagation import Request
 from flask import current_app, signals
 # needed to build a request object from environ in the middleware
 from werkzeug.wrappers import Request
-from beeline.middleware.werkzeug import WerkzeugRequest
+from beeline.middleware.wsgi import WSGIRequest
 
 
 class HoneyMiddleware(object):
@@ -39,10 +39,9 @@ class HoneyWSGIMiddleware(object):
 
     def __call__(self, environ, start_response):
         req = Request(environ, shallow=True)
-        wr = WerkzeugRequest(environ)
+        wr = WSGIRequest("flask", environ)
 
-        root_span = beeline.propagate_and_start_trace(
-            self.get_context_from_environ, wr)
+        root_span = beeline.propagate_and_start_trace(wr.request_context(), wr)
 
         def _start_response(status, headers, *args):
             status_code = int(status[0:4])
@@ -55,26 +54,6 @@ class HoneyWSGIMiddleware(object):
             return start_response(status, headers, *args)
 
         return self.app(environ, _start_response)
-
-    def get_context_from_environ(self, environ):
-        request_method = environ.get('REQUEST_METHOD')
-        if request_method:
-            trace_name = "flask_http_%s" % request_method.lower()
-        else:
-            trace_name = "flask_http"
-
-        return {
-            "type": "http_server",
-            "name": trace_name,
-            "request.host": environ.get('HTTP_HOST'),
-            "request.method": request_method,
-            "request.path": environ.get('PATH_INFO'),
-            "request.remote_addr": environ.get('REMOTE_ADDR'),
-            "request.content_length": environ.get('CONTENT_LENGTH', 0),
-            "request.user_agent": environ.get('HTTP_USER_AGENT'),
-            "request.scheme": environ.get('wsgi.url_scheme'),
-            "request.query": environ.get('QUERY_STRING')
-        }
 
 
 class HoneyDBMiddleware(object):
